(fix|minor) Yaml configuration reader should check for file existence.
(fix|major) Virtual machine share statistics should be updated only if the VM is currently powered on; furthermore, the update of these statistics should be invoked every time a VM is powered-on and is resumed.

(fix|major) The computation of physical machine energy consumption should take care the idle energy spent over all the uptime.